Text.Regex.Applicative.Reference

Description

Reference implementation (using backtracking).

This is exposed for testing purposes only!

Documentation

reference :: RE s a -> [s] -> Maybe a

reference r s should give the same results as s =~ r.

However, this is not very efficient implementation and is supposed to be used for testing only.